backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
awscli==1.38.15
blinker==1.9.0
botocore==1.37.15
cachetools==6.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
dataclasses-json==0.6.7
demucs==4.0.1
Django==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
docutils==0.16
dora_search==0.1.12
einops==0.8.1
filelock==3.17.0
frozenlist==1.7.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
julius==0.2.7
lameenc==1.8.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.48.1
networkx==3.4.2
numpy==2.2.3
omegaconf==2.3.0
openunmix==1.3.0
orjson==3.10.18
packaging==24.2
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rpds-py==0.26.0
rsa==4.7.2
s3transfer==0.11.4
safetensors==0.5.3
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
streamlit==1.47.1
stripe==7.0.0
submitit==1.5.2
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.21.2
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.1
treetable==0.2.5
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.45.1
whitenoise==6.9.0
yarl==1.20.1
zstandard==0.23.0
